Prac13_WebSocket/requirements.txt
aiomysql==0.2.0
annotated-types==0.6.0
anyio==4.3.0
asyncio==3.4.3
click==8.1.7
colorama==0.4.6
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
pydantic==2.6.4
pydantic_core==2.16.3
PyMySQL==1.1.0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
websockets==12.0
